1.1.2 Fix issues with missing timezone During setup a warning is displayed if there is no date.timezone. It is still possible to continue setup without changing the settings
The hook is executed before the application controllers and sets date_default_timezone_set to date_default_timezone_get which returns UTC if there is no timezone set in ini, via ini_set or in server configuration (available to php). With ENVIRONMENT set to production a warning still appears in log.